RDK-36812 : RPI Fix RDKMEDIAPLAYER build error in 64-bit
Reason for change: unsigned int to size_t error

| ../git/rdkmediaplayer.cpp:764:42: error: cast from 'RDKMediaPlayerImpl*' to 'uint32_t' {aka 'unsigned int'} loses precision [-fpermissive]
|   764 |         resp = (std::to_string((uint32_t)m_pImpl)).c_str();
|       |                                          ^~~~~~~
| ../git/rdkmediaplayer.cpp:785:51: error: cast from 'RDKMediaPlayerImpl*' to 'uint32_t' {aka 'unsigned int'} loses precision [-fpermissive]
|   785 |                 resp = (std::to_string((uint32_t)*it)).c_str();

Test Procedure: Build Images and do basic sanity
Risks: Low
(cherry picked from commit e6c805cc56ded258db6e33b69fda30d2a319900b)
(cherry picked from commit f54df176ec7674e7e3eed1f407de8b56673510ce)

+++ rdkmediaplayer.cpp
@@ -761,7 +761,7 @@ rtError RDKMediaPlayer::open(rtString openData, rtString resp)
         {
             LOG_INFO("CAS Managment session is already avialable");
         }
-        resp = (std::to_string((uint32_t)m_pImpl)).c_str();
+        resp = (std::to_string((uintptr_t)m_pImpl)).c_str();
         LOG_INFO("resp = %x", resp.cString());
     }
     // For Live Playback Session
@@ -782,7 +782,7 @@ rtError RDKMediaPlayer::open(rtString openData, rtString resp)
             if( !((*it)->isManagementSession()) && (*it)->doCanPlayURL(m_currentURL) )
             {
                 LOG_INFO("Reusing cached player");
-                resp = (std::to_string((uint32_t)*it)).c_str();
+                resp = (std::to_string((uintptr_t)*it)).c_str();
                 LOG_INFO("resp = %x", resp.cString());
             }
         }
